Changed a couple more ways values were stored.

+++ player/move_gen.c
@@ -156,18 +156,18 @@ int square_to_str(square_t sq, char *buf, size_t bufsize) {
 }
 
 // direction map
-static int dir[8] = { -ARR_WIDTH - 1, -ARR_WIDTH, -ARR_WIDTH + 1, -1, 1,
+static int16_t dir[8] = { -ARR_WIDTH - 1, -ARR_WIDTH, -ARR_WIDTH + 1, -1, 1,
                       ARR_WIDTH - 1, ARR_WIDTH, ARR_WIDTH + 1 };
-int dir_of(int i) {
+int16_t dir_of(int i) {
   tbassert(i >= 0 && i < 8, "i: %d\n", i);
   return dir[i];
 }
 
 
 // directions for laser: NN, EE, SS, WW
-static int beam[NUM_ORI] = {1, ARR_WIDTH, -1, -ARR_WIDTH};
+static int16_t beam[NUM_ORI] = {1, ARR_WIDTH, -1, -ARR_WIDTH};
 
-int beam_of(int direction) {
+int16_t beam_of(int direction) {
   tbassert(direction >= 0 && direction < NUM_ORI, "dir: %d\n", direction);
   return beam[direction];
 }
